Organize MobileNode receivers
